Remove not-owned plots from generic view Fixes DP-5763
When no owner type is selected (and consequently all owner types are
selected) it doesn't make sense to show not-owned plots (with owned
apartment rights for the owner type).

This is achieved by omitting the step that added those to the details
from the sql batch processing.

It also has a performance benefit.