Search Polygons instead of Points Spike for DP-5732
Geoselection on the map finds all plots that intersect with the given
shape when searching, the list views however only return those plots
whose centroids fall within the shape. This means that the search
results don't match.

This commit changes the geometry in the doc for elastic from point to
shape, and uses the predefined field geo_polygon.