Index 'appartement' on plot, not point
'Appartementen' are indexed with their point, and therefor didn't show
up in shape selections where the plot intersected with the shape
selection but the shape did not encompas the points.

Drawing on the map did select them.

To make this uniform over selection mechanisms, the 'appartementen' are
indexed with their plots, not their points.