DP-6461 - Get rid of errors reported in Sentry about GDAL_ERROR 10: b"Pointer 'hObject' is NULL in 'GDALGetDescription'.
In Sentry the following errors were reported.

GDAL_ERROR 4: b'SRID=4326;POLYGON ((4.9090864 52.3569284, 4.9098255 52.3558518, 4.9112653 52.3561219, 4.9121873 52.3563823, 4.9113125 52.3574697, 4.9090864 52.3569284)): No such file or directory'

and

GDAL_ERROR 10: b"Pointer 'hObject' is NULL in 'GDALGetDescription'.

These reported errors are more like warnings because it is working correctly. But they happen
because the query parameter is converted to Polygon in a early stage whewre it is later converted to
String again. When the lowlevel filter wants to convert it to a Geometry again it first fails
in get_raster_prep_value with :

venv/lib/python3.7/site-packages/django/contrib/gis/db/models/fields.py:182

There it generates these nasty error messages.

Ths can be fixed by converting the parameter to polygon in a later stage,
just before the actual filtering.