DP-6827 - Remove condition to select first kadastraal object in 'lijst' of kadastrale objecten.
This does not work correctly when extra conditions are added such as  eigenaar_type=Pandeigenaar
All references to kadastraal_object_index are removed, as well as the redis instance to
create these indexes.

Duplicates of kadastrale objects in the 'lijst' of kadastrale objects are now removed with
python code, and not with  elastic search. This means that the total_counts are not correct,
and the pagesizes are not 100 but less.