BRK dataselectie result fix
Currently after the query result, the kadastraal objects are post filtered on uniqueness. Due that multiple addresses can be related to a kadastraal object, the kadastraal object can appear more then once in the result set. To cope with this, kadastraal objects are removed after the query result. However, this causes a mismatch between the size URL parameter and the final result. Hence, the final result can be smaller (due to te removal of objects) then the given size value. This can be confusing for the user.

Now, after the de-duplication of the kadastraal objects, the number of results is finally defined by the given size value. Making the numer of results equal to the size parameter.